[Beginner] Package Refactoring
- Rename app package name.
- Rename app name from My Application to BMS Clone.
