+++ settings.gradle
@@ -1,2 +1,2 @@
 include ':app'
-rootProject.name = "My Application"
+rootProject.name = "BMS Clone"
